Move bin script to hidden CLI command
This makes it easier for anyone to install and run with, even if they
haven't set up the repo for local development.
