+++ packages/cli/oclif.manifest.json
